♻️ refactor(document): relocate headlessEditor to neutral document service location
Move headlessEditor from agentDocuments/ to document/ so PersonalPagesService
can import it as a peer without creating a personal→agent dependency.
