+++ src/server/services/agentSignal/policies/analyzeIntent/actions/__tests__/skillManagement.test.ts
@@ -27,7 +27,7 @@ const skillMaintainerService = {
 };
 const createSkill = vi.fn();
 
-vi.mock('@/server/services/agentDocuments/headlessEditor', () => ({
+vi.mock('@/server/services/document/headlessEditor', () => ({
   createMarkdownEditorSnapshot: vi.fn(async (content: string) => ({
     content,
     editorData: { markdown: content },
