+++ src/server/services/document/headlessEditor.ts
